+++ src/console/logs/logs.v
@@ -12,7 +12,7 @@ struct Config {
 	api_key string [required]
 }
 
-// cmd returns the cli module that handles the build repos API.
+// cmd returns the cli module that handles the build logs API.
 pub fn cmd() cli.Command {
 	return cli.Command{
 		name: 'logs'
@@ -33,8 +33,8 @@ pub fn cmd() cli.Command {
 						flag: cli.FlagType.int
 					},
 					cli.Flag{
-						name: 'repo'
-						description: 'Only return logs for this repo id.'
+						name: 'target'
+						description: 'Only return logs for this target id.'
 						flag: cli.FlagType.int
 					},
 					cli.Flag{
@@ -79,9 +79,9 @@ pub fn cmd() cli.Command {
 						filter.offset = u64(offset)
 					}
 
-					repo_id := cmd.flags.get_int('repo')?
-					if repo_id != 0 {
-						filter.repo = repo_id
+					target_id := cmd.flags.get_int('target')?
+					if target_id != 0 {
+						filter.target = target_id
 					}
 
 					tz_offset := time.offset()
@@ -168,10 +168,10 @@ pub fn cmd() cli.Command {
 
 // print_log_list prints a list of logs.
 fn print_log_list(logs []BuildLog) ? {
-	data := logs.map([it.id.str(), it.repo_id.str(), it.start_time.local().str(),
+	data := logs.map([it.id.str(), it.target_id.str(), it.start_time.local().str(),
 		it.exit_code.str()])
 
-	println(console.pretty_table(['id', 'repo', 'start time', 'exit code'], data)?)
+	println(console.pretty_table(['id', 'target', 'start time', 'exit code'], data)?)
 }
 
 // list prints a list of all build logs.
@@ -182,10 +182,10 @@ fn list(conf Config, filter BuildLogFilter) ? {
 	print_log_list(logs)?
 }
 
-// list prints a list of all build logs for a given repo.
-fn list_for_repo(conf Config, repo_id int) ? {
+// list prints a list of all build logs for a given target.
+fn list_for_target(conf Config, target_id int) ? {
 	c := client.new(conf.address, conf.api_key)
-	logs := c.get_build_logs_for_repo(repo_id)?.data
+	logs := c.get_build_logs_for_target(target_id)?.data
 
 	print_log_list(logs)?
 }