+++ src/console/targets/build.v
@@ -1,14 +1,14 @@
-module git
+module targets
 
 import client
 import docker
 import os
 import build
 
-// build builds every Git repo in the server's list.
+// build locally builds the target with the given id.
 fn build(conf Config, repo_id int) ? {
 	c := client.new(conf.address, conf.api_key)
-	repo := c.get_git_repo(repo_id)?
+	repo := c.get_target(repo_id)?
 
 	build_arch := os.uname().machine