+++ src/db/db.v
@@ -13,8 +13,12 @@ struct MigrationVersion {
 }
 
 const (
-	migrations_up   = [$embed_file('migrations/001-initial/up.sql')]
-	migrations_down = [$embed_file('migrations/001-initial/down.sql')]
+	migrations_up   = [
+		$embed_file('migrations/001-initial/up.sql'),
+		$embed_file('migrations/002-rename-to-targets/up.sql'),
+	]
+	migrations_down = [$embed_file('migrations/001-initial/down.sql'),
+		$embed_file('migrations/002-rename-to-targets/down.sql')]
 )
 
 // init initializes a database & adds the correct tables.