+++ src/models/logs.v
@@ -5,7 +5,7 @@ import time
 pub struct BuildLog {
 pub mut:
 	id         int       [primary; sql: serial]
-	repo_id    int       [nonull]
+	target_id  int       [nonull]
 	start_time time.Time [nonull]
 	end_time   time.Time [nonull]
 	arch       string    [nonull]
@@ -16,7 +16,7 @@ pub mut:
 pub fn (bl &BuildLog) str() string {
 	mut parts := [
 		'id: $bl.id',
-		'repo id: $bl.repo_id',
+		'target id: $bl.target_id',
 		'start time: $bl.start_time.local()',
 		'end time: $bl.end_time.local()',
 		'duration: ${bl.end_time - bl.start_time}',
@@ -33,7 +33,7 @@ pub struct BuildLogFilter {
 pub mut:
 	limit      u64 = 25
 	offset     u64
-	repo       int
+	target     int
 	before     time.Time
 	after      time.Time
 	arch       string