+++ src/models/targets.v
@@ -1,18 +1,18 @@
 module models
 
-pub struct GitRepoArch {
+pub struct TargetArch {
 pub:
-	id      int    [primary; sql: serial]
-	repo_id int    [nonull]
-	value   string [nonull]
+	id        int    [primary; sql: serial]
+	target_id int    [nonull]
+	value     string [nonull]
 }
 
 // str returns a string representation.
-pub fn (gra &GitRepoArch) str() string {
+pub fn (gra &TargetArch) str() string {
 	return gra.value
 }
 
-pub struct GitRepo {
+pub struct Target {
 pub mut:
 	id int [primary; sql: serial]
 	// URL of the Git repository
@@ -25,11 +25,11 @@ pub mut:
 	schedule string
 	// On which architectures the package is allowed to be built. In reality,
 	// this controls which builders will periodically build the image.
-	arch []GitRepoArch [fkey: 'repo_id']
+	arch []TargetArch [fkey: 'target_id']
 }
 
 // str returns a string representation.
-pub fn (gr &GitRepo) str() string {
+pub fn (gr &Target) str() string {
 	mut parts := [
 		'id: $gr.id',
 		'url: $gr.url',
@@ -44,7 +44,7 @@ pub fn (gr &GitRepo) str() string {
 }
 
 [params]
-pub struct GitRepoFilter {
+pub struct TargetFilter {
 pub mut:
 	limit  u64 = 25
 	offset u64